Build container from Docker directory
This will reduce the amount of context that needs to be sent to Docker
when building.  The current directory could have a bunch of source code,
built binaries, etc. that are used for the current project so it's
faster to omit it.  Any files that need to be installed into the
container should be placed in the Docker directory or below.