Build NeoVim from source.
NeoVim is compiled in a single layer

Instead of cloning in one layer and compiling in the next, do it all in one
layer.  Having two layers is useful for debugging because it avoids
re-downloading the entire repo while debugging the `Dockerfile`.  However,
it's just a waste once it has stabilized.

In addition, the repo is deleted in the same layer.  This reduces the size of
the image by approximately 50%.